# documentation/openapi/metrics.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Metrics HTTP API",
    "description": "Prometheus-compatible query and ingestion API. Configure path_prefix to prepend a common prefix to data routes.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/read/ns/{namespace}/api/v1/label/{name}/values": {
      "get": {
        "tags": [
          "metadata"
        ],
        "operationId": "label_values",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Label name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "match[]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Label values",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/read/ns/{namespace}/api/v1/metadata": {
      "get": {
        "tags": [
          "metadata"
        ],
        "operationId": "metadata",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metric metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/read/ns/{namespace}/federate": {
      "get": {
        "tags": [
          "query"
        ],
        "operationId": "federate",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "match[]",
            "in": "query",
            "description": "Series selectors",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prometheus text exposition",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/write/ns/{namespace}/api/v1/write": {
      "post": {
        "tags": [
          "ingest"
        ],
        "operationId": "remote_write",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Snappy-compressed Prometheus remote-write request: `prometheus.WriteRequest` (1.0, default) or `io.prometheus.write.v2.Request` when the content type carries `proto=io.prometheus.write.v2.Request`. Both carry native histograms.",
          "content": {
            "application/x-protobuf": {
              "schema": {
                "$ref": "#/components/schemas/BinaryBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Write accepted; 2.0 requests also receive X-Prometheus-Remote-Write-{Samples,Histograms,Exemplars}-Written headers"
          },
          "400": {
            "description": "Invalid remote-write request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported remote-write protobuf message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object"
          },
          "status": {
            "type": "string",
            "description": "Prometheus response status (`success` or `error`)."
          }
        }
      },
      "BinaryBody": {
        "type": "string",
        "format": "binary"
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "status",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "error_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}